+++ integrations/layer/index.js
@@ -4,31 +4,42 @@ var logger = require('../../services/logger');
 var Routes = require('./routes');
 var Setup = require('./setup');
 
-function Checker(opts) {
+function Checker(opts, Implementation) {
   var integrationValid = false;
 
-  function hasLayerIntegration() {
+  function hasIntegration() {
     return opts.integrations && opts.integrations.layer;
   }
 
-  function isLayerProperlyIntegrated() {
+  function isProperlyIntegrated() {
     return opts.integrations.layer.serverApiToken &&
       opts.integrations.layer.appId;
   }
 
-  if (hasLayerIntegration()) {
-    if (isLayerProperlyIntegrated()) {
-      integrationValid = true;
-    } else {
-      logger.error('Cannot setup properly your Layer integration.');
+  function isMappingValid() {
+    var models = Implementation.getModels();
+    var mappingValid = true;
+    _.map(opts.integrations.layer.mapping, function (mappingValue) {
+      var collectionName = mappingValue.split('.')[0];
+      if (!models[collectionName]) {
+        mappingValid = false;
+      }
+    });
+
+    if (!mappingValid) {
+      logger.error('Cannot find some Layer integration mapping values (' +
+        opts.integrations.layer.mapping + ') among the project models:\n' +
+        _.keys(models).join(', '));
     }
+
+    return mappingValid;
   }
 
   function castToArray(value) {
     return _.isString(value) ? [value] : value;
   }
 
-  function integrationCollectionMatch(Implementation, integration, model) {
+  function integrationCollectionMatch(integration, model) {
     if (!integrationValid) { return; }
 
     var models = Implementation.getModels();
@@ -45,26 +56,25 @@ function Checker(opts) {
       Implementation.getModelName(model)) > -1;
   }
 
-  if (hasLayerIntegration()) {
-    if (isLayerProperlyIntegrated()) {
+  if (hasIntegration()) {
+    if (isProperlyIntegrated()) {
       opts.integrations.layer.mapping =
         castToArray(opts.integrations.layer.mapping);
-      integrationValid = true;
+      integrationValid = isMappingValid();
     } else {
       logger.error('Cannot setup properly your Layer integration.');
     }
   }
 
-  this.defineRoutes = function (app, model, Implementation) {
+  this.defineRoutes = function (app, model) {
     if (!integrationValid) { return; }
 
-    if (integrationCollectionMatch(Implementation, opts.integrations.layer,
-      model)) {
+    if (integrationCollectionMatch(opts.integrations.layer, model)) {
       new Routes(app, model, Implementation, opts).perform();
     }
   };
 
-  this.defineCollections = function (Implementation, collections) {
+  this.defineCollections = function (collections) {
     if (!integrationValid) { return; }
 
     _.each(opts.integrations.layer.mapping,
@@ -74,17 +84,15 @@ function Checker(opts) {
       });
   };
 
-  this.defineFields = function (Implementation, model, schema) {
+  this.defineFields = function (model, schema) {
     if (!integrationValid) { return; }
 
-    if (integrationCollectionMatch(Implementation, opts.integrations.layer,
-      model)) {
+    if (integrationCollectionMatch(opts.integrations.layer, model)) {
         Setup.createFields(Implementation, model, schema.fields);
     }
   };
 
-  this.defineSerializationOption = function (Implementation, model, schema,
-    dest, field) {
+  this.defineSerializationOption = function (model, schema, dest, field) {
     if (integrationValid && field.integration === 'layer') {
       dest[field.field] = {
         ref: 'id',
@@ -107,4 +115,3 @@ function Checker(opts) {
 }
 
 module.exports = Checker;
-