+++ integrations/mixpanel/index.js
@@ -1,9 +1,10 @@
 'use strict';
+var _ = require('lodash');
 var logger = require('../../services/logger');
 var Setup = require('./setup');
 var Routes = require('./routes');
 
-function Checker(opts) {
+function Checker(opts, Implementation) {
   var integrationValid = false;
 
   function hasIntegration() {
@@ -16,7 +17,21 @@ function Checker(opts) {
       opts.integrations.mixpanel.mapping && opts.integrations.mixpanel.mixpanel;
   }
 
-  function integrationCollectionMatch(Implementation, integration, model) {
+  function isMappingValid() {
+    var models = Implementation.getModels();
+    var collectionName = opts.integrations.mixpanel.mapping.split('.')[0];
+    var mappingValid = !!models[collectionName];
+
+    if (!mappingValid) {
+      logger.error('Cannot find some Mixpanel integration mapping values (' +
+        opts.integrations.mixpanel.mapping + ') among the project models:\n' +
+        _.keys(models).join(', '));
+    }
+
+    return mappingValid;
+  }
+
+  function integrationCollectionMatch(integration, model) {
     if (!integrationValid) { return; }
 
     var models = Implementation.getModels();
@@ -25,48 +40,44 @@ function Checker(opts) {
   }
 
   if (hasIntegration()) {
-    if (isProperlyIntegrated()) {
+    if (isProperlyIntegrated() && isMappingValid()) {
       integrationValid = true;
     } else {
       logger.error('Cannot setup properly your Mixpanel integration.');
     }
   }
 
-  this.defineRoutes = function (app, model, Implementation) {
+  this.defineRoutes = function (app, model) {
     if (!integrationValid) { return; }
 
-    if (integrationCollectionMatch(Implementation, opts.integrations.mixpanel,
-      model)) {
+    if (integrationCollectionMatch(opts.integrations.mixpanel, model)) {
       new Routes(app, model, Implementation, opts).perform();
     }
   };
 
-  this.defineCollections = function (Implementation, model, schema) {
+  this.defineCollections = function (model, schema) {
     if (!integrationValid) { return; }
 
     Setup.createCollections(Implementation, model, schema, opts);
   };
 
-  this.defineSegments = function (Implementation, model, schema) {
+  this.defineSegments = function (model, schema) {
     if (!integrationValid) { return; }
 
-    if (integrationCollectionMatch(Implementation, opts.integrations.mixpanel,
-      model)) {
+    if (integrationCollectionMatch(opts.integrations.mixpanel, model)) {
       Setup.createSegments(Implementation, model, schema, opts);
     }
   };
 
-  this.defineFields = function (Implementation, model, schema) {
+  this.defineFields = function (model, schema) {
     if (!integrationValid) { return; }
 
-    if (integrationCollectionMatch(Implementation, opts.integrations.mixpanel,
-      model)) {
+    if (integrationCollectionMatch(opts.integrations.mixpanel, model)) {
       Setup.createFields(Implementation, model, schema.fields);
     }
   };
 
-  this.defineSerializationOption = function (Implementation, model, schema,
-    dest, field) {
+  this.defineSerializationOption = function (model, schema, dest, field) {
     if (integrationValid && field.integration === 'mixpanel') {
       dest[field.field] = {
         ref: 'id',