+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Exports - Escape special characters for the string fields. 
+- Integrations - Display models "mapping" errors if any.
 
 ## RELEASE 1.2.2 - 2017-08-24
 ### Changed